Strip timezone information from date fields
Currently, timezone information for default values on date fields is left
as-is. More often than not, this default value contains timezone information.

This commit strips the timezone information from the default value because
it is not possible to provide timezone information in a MySQL date field.
Proper conversion is, as always with MySQL, left to the application.